tests/e2e/mock-oidc/requirements.txt
cryptography>=42.0.0
PyJWT>=2.8.0
